+++ packages/chatbot-ui/src/chatbot-ui/components/MessageArea.tsx
@@ -1,7 +1,7 @@
 import { Loader2 } from 'lucide-react'
 import type { ReactNode } from 'react'
-import { Avatar, AvatarFallback, AvatarImage } from '#/components/ui/avatar'
-import { Card, CardContent } from '#/components/ui/card'
+import { Avatar, AvatarFallback, AvatarImage } from '#/chatbot-ui/components/ui/avatar'
+import { Card, CardContent } from '#/chatbot-ui/components/ui/card'
 
 // メッセージの型定義
 export type Message = {
