+++ packages/chatbot-ui/src/chatbot-ui/components/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 import { Moon, Sun } from 'lucide-react'
 
-import { Button } from '#/components/ui/button'
-import { useTheme } from '#/hooks/useTheme'
+import { Button } from '#/chatbot-ui/components/ui/button'
+import { useTheme } from '#/chatbot-ui/hooks/useTheme'
 
 interface ThemeToggleProps {
   showLabel?: boolean
