+++ packages/chatbot-ui/src/chatbot-ui/components/ui/lib/utils.ts
