+++ packages/chatbot-ui/src/chatbot-ui/components/ui/separator.tsx
@@ -1,7 +1,7 @@
 import * as SeparatorPrimitive from '@radix-ui/react-separator'
 import type * as React from 'react'
 
-import { cn } from '#/components/ui/lib/utils'
+import { cn } from '#/chatbot-ui/components/ui/lib/utils'
 
 function Separator({
   className,
