+++ packages/chatbot-ui/src/chatbot-ui/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useChat'
+export * from './useTheme'
