+++ packages/chatbot-ui/src/chatbot-ui/hooks/useTheme.ts
@@ -57,3 +57,17 @@ export function useTheme() {
     setTheme: setThemeAndStore,
   }
 }
+
+export function initializeTheme() {
+  const savedTheme = localStorage.getItem('theme') || 'system'
+  const root = document.documentElement
+
+  root.classList.remove('light', 'dark')
+
+  if (savedTheme === 'system') {
+    const systemTheme = window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light'
+    root.classList.add(systemTheme)
+  } else {
+    root.classList.add(savedTheme)
+  }
+}
